style&test: Sections 클래스 메서드 순서 변경 및 라인관련 인수테스트 분리
